diffusion_tutorial/requirements.txt
datasets==3.0.1
opencv-python-headless==4.10.0.84
Pillow==9.4.0
diffusers==0.30.3
torch==2.1.0
tqdm==4.66.2